requirements.txt
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
sqlparse==0.4.2
django-simpleui==2023.3.1
django-import-export==2.0.1
requests
pycryptodome